+++ apps/www/src/lib/components/docs/page-header/index.ts
@@ -1,13 +1,16 @@
 import Root from "./page-header.svelte";
 import Heading from "./page-header-heading.svelte";
 import Description from "./page-header-description.svelte";
+import Actions from "./page-header-actions.svelte";
 
 export {
 	Root,
 	Heading,
 	Description,
+	Actions,
 	//
 	Root as PageHeader,
 	Heading as PageHeaderHeading,
 	Description as PageHeaderDescription,
+	Actions as PageHeaderActions,
 };