+++ apps/www/src/routes/(blocks)/blocks/[style]/[name]/+page.server.ts
@@ -1,0 +1,36 @@
+import { getAllBlockIds, isDemo } from "$lib/blocks.js";
+import { blockMeta } from "$lib/config/blocks.js";
+import { isStyle, styles } from "$lib/registry/styles.js";
+import { error } from "@sveltejs/kit";
+import type { EntryGenerator } from "./$types.js";
+
+export const prerender = true;
+
+export const entries: EntryGenerator = async () => {
+	console.log("Prerendering /blocks/[style]/[name]");
+	const blockIds = await getAllBlockIds();
+	const entries = styles.flatMap((style) => {
+		return blockIds.map((name) => ({ name, style: style.name }));
+	});
+
+	return entries;
+};
+
+export const load = async (event) => {
+	const { name, style } = event.params;
+	if (!isStyle(style)) error(404, "Style not found");
+	if (!isDemo(name)) error(404, "Block not found");
+
+	const { className, iframeHeight } = blockMeta[style][name];
+
+	return {
+		block: {
+			name: name,
+			style: style,
+			container: {
+				height: iframeHeight,
+				className,
+			},
+		},
+	};
+};