+++ packages/cli/src/utils/transformers.ts
@@ -1,11 +1,4 @@
-import { Config } from "./get-config.js";
-
-export type TransformOpts = {
-	filename: string;
-	content: string;
-	config: Config;
-	// baseColor?: z.infer<typeof registryBaseColorSchema>; - will use later
-};
+import type { Config } from "./get-config.js";
 
 export function transformImports(content: string, config: Config) {
 	let str = content.replace(/\$lib\/registry\/[^/]+/g, config.aliases.components);